requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
chromadb==0.4.22
python-dotenv==1.0.0
python-dateutil==2.8.2
numpy<2.0 